+++ source/components/__tests__/Form.test.js
@@ -44,7 +44,6 @@ describe('components / Form', () => {
             'registerField',
             'unregisterField',
             'resetForm',
-            'updateForm',
           ]),
         );
 
@@ -82,7 +81,7 @@ describe('components / Form', () => {
       createElement(
         Form,
         { path: 'form' },
-        createElement(Field, { name: 'test', component: 'input' }),
+        createElement(Field, { name: 'test', component: 'input', changed: true, touched: true }),
       ),
       global.formContext,
     );
@@ -90,9 +89,9 @@ describe('components / Form', () => {
     expect(global.store.getState().form).toEqual({
       fields: {
         test: {
-          changed: false,
+          changed: true,
           disabled: false,
-          touched: false,
+          touched: true,
           errors: [],
           valid: true,
           value: '',
@@ -100,8 +99,8 @@ describe('components / Form', () => {
       },
       submitted: false,
       submitting: false,
-      touched: false,
-      changed: false,
+      touched: true,
+      changed: true,
       valid: true,
     });
   });