+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.4.1] - 2018-10-22
+
+- Usage of polyfills.
+
+## [1.4.0] - 2018-10-20
+
+- Support to reset form to empty or initial state.
+
+## [1.3.7] - 2018-10-20
+
+- If some field was touched or changed then form was touched or changed too.
+
+## [1.3.6] - 2018-10-18
+
+- Add fields in the store after Form mount. (bugfix)
+
 ## [1.3.5] - 2018-10-13
 
 - Update components to new react api.